Pull out header unpacking and parsing logic into dedicated submodules
